chore: Move component directory
